Correction du constructeur par copie de Vector

+++ src/Point.cpp
@@ -26,7 +26,7 @@ void		Point::setPos(int x, int y)
 
 void		Point::calcRealpos(void)
 {
-  this->_pos.x = this->_scale / 2 + this->_x * this->_scale;
-  this->_pos.y = this->_scale / 2;
-  this->_pos.z = this->_scale / 2 + this->_y * this->_scale;
+  this->_pos.x = this->_scale / 2.0f + static_cast<double>(this->_x) * this->_scale;
+  this->_pos.y = this->_scale / 2.0f;
+  this->_pos.z = this->_scale / 2.0f + static_cast<double>(this->_y) * this->_scale;
 }
